+++ korge/src/commonMain/kotlin/korlibs/korge/view/ViewExt.kt
@@ -1,14 +1,13 @@
 package korlibs.korge.view
 
-import korlibs.time.Frequency
-import korlibs.time.timesPerSecond
-import korlibs.io.lang.Cancellable
+import korlibs.io.lang.*
 import korlibs.math.geom.*
+import korlibs.time.*
 
 val Double.fps get() = this.timesPerSecond
 val Int.fps get() = this.timesPerSecond
 
-fun <T : View> T.addUpdater(referenceFps: Frequency, first: Boolean = true, updatable: T.(scale: Double) -> Unit): Cancellable {
+fun <T : View> T.addUpdater(referenceFps: Frequency, first: Boolean = true, updatable: T.(scale: Float) -> Unit): Cancellable {
     val time = referenceFps.timeSpan
     return addUpdater(first) {
         updatable(it / time)
